// src/shared/api/models/getApiGiftCardsId200HistoryItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiGiftCardsId200HistoryItemInvoiceNumber } from './getApiGiftCardsId200HistoryItemInvoiceNumber';

export type GetApiGiftCardsId200HistoryItem = {
    salesId: string;
    invoiceNumber: GetApiGiftCardsId200HistoryItemInvoiceNumber;
    amountUsed: number;
    dateTime: string;
};