// src/shared/api/models/patchApiGiftCardsIdBody.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */

export type PatchApiGiftCardsIdBody = {
    /** @nullable */
    customerId?: number | null;
    /** @minimum 0 */
    originalValue?: number;
    /** @minimum 0 */
    residueValue?: number;
    /** @nullable */
    expiryDate?: string | null;
    /** @nullable */
    recipientName?: string | null;
    /** @nullable */
    note?: string | null;
};